:recycle: Logic Seperation
seperate logic from the DOM
